// src/app/character/enums/character.enum.ts
export const enum URLType {
    Comiclink = "comiclink",
    Detail = "detail",
    Wiki = "wiki",
}

export const enum Extension {
    GIF = "gif",
    Jpg = "jpg",
}

export const enum ItemType {
    Cover = "cover",
    Empty = "",
    InteriorStory = "interiorStory",
}

export const enum ImageAspect { 
    portrait = "/portrait_",
    standard = "/standard_",
    landscape = "/landscape_"
}

export const enum ImageSize { 
    small = "small.",
    medium = "medium.",
    xlarge = "xlarge.",
    fantastic = "fantastic.",
    uncanny = "uncanny.",
    incredible = "incredible.",
    amazing = "amazing."
}